Replace previous MarkerCluster implementation with new
