--- grommet-leaflet/src/layers/MarkerCluster2/MarkerCluster2.jsx
@@ -1,116 +1,0 @@
-import React from 'react';
-import ReactDOMServer from 'react-dom/server';
-import { createRoot } from 'react-dom/client';
-import { ThemeContext } from 'styled-components';
-import { v4 as uuidv4 } from 'uuid';
-import {
-  createElementObject,
-  createLayerComponent,
-  extendContext,
-} from '@react-leaflet/core';
-import L from 'leaflet';
-import 'leaflet.markercluster';
-import { Cluster, Popup } from '..';
-
-function createMarkerClusterGroup(props, context) {
-  const clusterProps = {};
-  const clusterEvents = {};
-
-  Object.entries(props).forEach(([key, value]) => {
-    if (key.startsWith('on')) {
-      clusterEvents[key] = value;
-    } else {
-      clusterProps[key] = value;
-    }
-  });
-
-  const markers = new L.MarkerClusterGroup({
-    showCoverageOnHover: false,
-    zoomToBoundsOnClick: false,
-    ...clusterProps,
-  });
-
-  // Bind events to markers
-  Object.entries(clusterEvents).forEach(([key, value]) => {
-    const event = `cluster${key.substring(2).toLowerCase()}`;
-    markers.on(event, value);
-  });
-
-  markers.on('clusterclick clusterkeypress', e => {
-    const { propagatedFrom } = e;
-    const popupContent = clusterProps.popup({ markers });
-    const popupId = `grommet-leaflet-popup-${uuidv4()}`;
-    const popup = new L.Popup();
-    // In order to take advantage of Leaflet's automatic popup placement
-    // and panning we first need to render the popup's contents statically.
-    // This establishes the popup's dimensions so that the map will pan if the
-    // marker is close to the map's bounds.
-    popup
-      .setLatLng(propagatedFrom.getLatLng())
-      .setContent(
-        `<div id="${popupId}">${ReactDOMServer.renderToString(
-          popupContent,
-        )}</div>`,
-      )
-      .openOn(context.map);
-
-    // We then render the popup's contents dynamically and replace the static
-    // content with an interactive ReactNode.
-    const domNode = document.getElementById(popupId);
-    const root = createRoot(domNode);
-    root.render(popupContent);
-  });
-
-  return createElementObject(
-    markers,
-    extendContext(context, {
-      layerContainer: markers,
-    }),
-  );
-}
-
-const LeafletMarkerCluster = createLayerComponent(createMarkerClusterGroup);
-
-const MarkerCluster2 = props => {
-  const { children, icon: iconProp, popup: popupProp, ...rest } = props;
-  const theme = React.useContext(ThemeContext);
-  const popupRef = React.useRef(null);
-
-  const createIcon = cluster => {
-    return L.divIcon({
-      // 'grommet-cluster-group' class prevents
-      // leaflet default divIcon styles
-      className: 'grommet-cluster-group',
-      html: ReactDOMServer.renderToString(
-        <ThemeContext.Provider value={theme}>
-          {iconProp ? (
-            React.cloneElement(iconProp({ cluster }), {
-              cluster,
-            })
-          ) : (
-            <Cluster cluster={cluster} />
-          )}
-        </ThemeContext.Provider>,
-      ),
-    });
-  };
-
-  const popup = ({ cluster }) =>
-    popupProp && typeof popupProp === 'function' ? (
-      <ThemeContext.Provider ref={popupRef} value={theme}>
-        <Popup>{popupProp({ cluster })}</Popup>
-      </ThemeContext.Provider>
-    ) : undefined;
-
-  return (
-    <LeafletMarkerCluster
-      iconCreateFunction={createIcon}
-      popup={popup}
-      {...rest}
-    >
-      {children}
-    </LeafletMarkerCluster>
-  );
-};
-
-export { MarkerCluster2 };
