--- grommet-leaflet/src/layers/MarkerCluster2/index.js
@@ -1,1 +1,0 @@
-export * from './MarkerCluster2';
