+++ grommet-leaflet/src/layers/index.js
@@ -3,7 +3,6 @@ export * from './Controls';
 export * from './Map';
 export * from './Marker';
 export * from './MarkerCluster';
-export * from './MarkerCluster2';
 export * from './Pin';
 export * from './Popup';
 export * from './TileLayer';
